+++ src/binding.js
@@ -4,7 +4,42 @@ function Binding () {
 
 var p = Binding.prototype
 
+/**
+ * Add a child binding to the tree.
+ *
+ * @param {String} key
+ * @param {Binding} child
+ */
+
 p.addChild = function (key, child) {
+  
+}
+
+/**
+ * Traverse along a path and trigger updates
+ * along the way.
+ *
+ * @param {String} path
+ */
+
+p.updatePath = function (path) {
+  
+}
+
+/**
+ * Trigger updates for the subtree starting at
+ * self as root.
+ */
+
+p.updateSubTree = function () {
+  
+}
+
+/**
+ * Notify all subscribers of it self
+ */
+
+p.notify = function () {
   
 }